feat: allow getting messages when not on chat page
also clean up and refactor some of the logic
