+++ js-peer/src/components/chat.tsx
@@ -3,158 +3,24 @@ import React, { useCallback, useEffect, useRef, useState } from 'react'
 import type { Message } from '@libp2p/interface'
 import { CHAT_FILE_TOPIC, CHAT_TOPIC, FILE_EXCHANGE_PROTOCOL } from '@/lib/constants'
 import { createIcon } from '@download/blockies'
-import { ChatMessage, useChatContext } from '../context/chat-ctx'
+import { ChatFile, ChatMessage, useChatContext } from '../context/chat-ctx'
 import { v4 as uuidv4 } from 'uuid';
-import { ChatFile, useFileChatContext } from '@/context/file-ctx'
 import { pipe } from 'it-pipe'
 import map from 'it-map'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
 import * as lp from 'it-length-prefixed'
+import { MessageComponent } from './message'
 
 interface MessageProps extends ChatMessage { }
 
-function Message({ msg, fileObjectUrl, from, peerId }: MessageProps) {
-  const msgref = React.useRef<HTMLLIElement>(null)
-  const { libp2p } = useLibp2pContext()
 
-
-  useEffect(() => {
-    const icon = createIcon({
-      seed: peerId,
-      size: 15,
-      scale: 3,
-    })
-    icon.className = 'rounded mr-2 max-h-10 max-w-10'
-    const childrenCount = msgref.current?.childElementCount
-    // Prevent inserting an icon more than once.
-    if (childrenCount && childrenCount < 2) {
-      msgref.current?.insertBefore(icon, msgref.current?.firstChild)
-    }
-  }, [peerId])
-
-  return (
-    <li ref={msgref} className={`flex ${from === 'me' ? 'justify-end' : 'justify-start'}`}>
-      <div
-
-        className="flex relative max-w-xl px-4 py-2 text-gray-700 rounded shadow bg-white"
-      >
-        <div className="block">
-          {msg}
-          <p>{fileObjectUrl ? <a href={fileObjectUrl} target="_blank"><b>Download</b></a> : ""}</p>
-          <p className="italic text-gray-400">{peerId !== libp2p.peerId.toString() ? `from: ${peerId.slice(-4)}` : null} </p>
-        </div>
-      </div>
-    </li>
-  )
-}
 
 export default function ChatContainer() {
   const { libp2p } = useLibp2pContext()
-  const { messageHistory, setMessageHistory } = useChatContext();
-  const { files, setFiles } = useFileChatContext();
+  const { messageHistory, setMessageHistory, files, setFiles } = useChatContext();
   const [input, setInput] = useState<string>('')
   const fileRef = useRef<HTMLInputElement>(null);
-
-  // Effect hook to subscribe to pubsub events and update the message state hook
-  useEffect(() => {
-    const messageCB = async (evt: CustomEvent<Message>) => {
-      console.log('gossipsub console log', evt.detail)
-      // FIXME: Why does 'from' not exist on type 'Message'?
-      const { topic, data } = evt.detail
-
-      switch (topic) {
-        case CHAT_TOPIC: {
-          chatMessageCB(evt, topic, data)
-          break
-        }
-        case CHAT_FILE_TOPIC: {
-          chatFileMessageCB(evt, topic, data)
-          break
-        }
-        default: {
-          throw new Error(`Unexpected gossipsub topic: ${topic}`)
-        }
-      }
-    }
-
-    const messageCBWrapper = (evt: Event) => {
-      const customEvent = evt as CustomEvent<Message>;
-      (async () => messageCB(customEvent))();
-    };
-
-    const chatMessageCB = (evt: CustomEvent<Message>, topic: string, data: Uint8Array) => {
-      const msg = new TextDecoder().decode(data)
-      console.log(`${topic}: ${msg}`)
-
-      // Append signed messages, otherwise discard
-      if (evt.detail.type === 'signed') {
-        setMessageHistory([...messageHistory, { msg, fileObjectUrl: undefined, from: 'other', peerId: evt.detail.from.toString() }])
-      }
-    }
-
-    const chatFileMessageCB = async (evt: CustomEvent<Message>, topic: string, data: Uint8Array) => {
-      const fileId = new TextDecoder().decode(data)
-
-      // if the message isn't signed, discard it.
-      if (evt.detail.type !== 'signed') {
-        return
-      }
-      const senderPeerId = evt.detail.from;
-
-      try {
-        const stream = await libp2p.dialProtocol(senderPeerId, FILE_EXCHANGE_PROTOCOL)
-        await pipe(
-          [uint8ArrayFromString(fileId)],
-          (source) => lp.encode(source),
-          stream,
-          (source) => lp.decode(source),
-          async function(source) {
-            for await (const data of source) {
-              const body: Uint8Array = data.subarray()
-              console.log(`request_response: response received: size:${body.length}`)
-
-              const msg: ChatMessage = {
-                msg: newChatFileMessage(fileId, body),
-                fileObjectUrl: window.URL.createObjectURL(new Blob([body])),
-                from: 'other',
-                peerId: senderPeerId.toString(),
-              }
-              setMessageHistory([...messageHistory, msg])
-            }
-          }
-        )
-      } catch (e) {
-        console.error(e)
-      }
-    }
-
-    libp2p.services.pubsub.addEventListener('message', messageCBWrapper)
-
-    libp2p.handle(FILE_EXCHANGE_PROTOCOL, ({ stream }) => {
-      pipe(
-        stream.source,
-        (source) => lp.decode(source),
-        (source) => map(source, async (msg) => {
-          const fileId = uint8ArrayToString(msg.subarray())
-          const file = files.get(fileId)!
-          return file.body
-        }),
-        (source) => lp.encode(source),
-        stream.sink,
-      )
-    })
-
-    return () => {
-      (async () => {
-        // Cleanup handlers 👇
-        // libp2p.services.pubsub.unsubscribe(CHAT_TOPIC)
-        // libp2p.services.pubsub.unsubscribe(CHAT_FILE_TOPIC)
-        libp2p.services.pubsub.removeEventListener('message', messageCBWrapper)
-        await libp2p.unhandle(FILE_EXCHANGE_PROTOCOL)
-      })();
-    }
-  }, [libp2p, messageHistory, setMessageHistory, files])
 
   const sendMessage = useCallback(async () => {
     if (input === '') return
@@ -284,7 +150,7 @@ export default function ChatContainer() {
               <ul className="space-y-2">
                 {/* messages start */}
                 {messageHistory.map(({ msg, fileObjectUrl, from, peerId }, idx) => (
-                  <Message key={idx} msg={msg} fileObjectUrl={fileObjectUrl} from={from} peerId={peerId} />
+                  <MessageComponent key={idx} msg={msg} fileObjectUrl={fileObjectUrl} from={from} peerId={peerId} />
                 ))}
                 {/* messages end */}
               </ul>
@@ -370,95 +236,3 @@ export default function ChatContainer() {
   )
 }
 
-export function RoomList() {
-  return (
-    <div className="border-r border-gray-300 lg:col-span-1">
-      <div className="mx-3 my-3">
-        <div className="relative text-gray-600">
-          <span className="absolute inset-y-0 left-0 flex items-center pl-2">
-            <svg
-              fill="none"
-              stroke="currentColor"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth="2"
-              viewBox="0 0 24 24"
-              className="w-6 h-6 text-gray-300"
-            >
-              <path d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"></path>
-            </svg>
-          </span>
-          <input
-            type="search"
-            className="block w-full py-2 pl-10 bg-gray-100 rounded outline-none"
-            name="search"
-            placeholder="Search"
-            required
-          />
-        </div>
-      </div>
-
-      <ul className="overflow-auto h-[32rem]">
-        <h2 className="my-2 mb-2 ml-2 text-lg text-gray-600">Chats</h2>
-        <li>
-          <a className="flex items-center px-3 py-2 text-sm transition duration-150 ease-in-out border-b border-gray-300 cursor-pointer hover:bg-gray-100 focus:outline-none">
-            <img
-              className="object-cover w-10 h-10 rounded-full"
-              src="https://github.com/2color.png"
-              alt="username"
-            />
-            <div className="w-full pb-2">
-              <div className="flex justify-between">
-                <span className="block ml-2 font-semibold text-gray-600">
-                  Daniel
-                </span>
-                <span className="block ml-2 text-sm text-gray-600">
-                  25 minutes
-                </span>
-              </div>
-              <span className="block ml-2 text-sm text-gray-600">bye</span>
-            </div>
-          </a>
-          <a className="flex items-center px-3 py-2 text-sm transition duration-150 ease-in-out bg-gray-100 border-b border-gray-300 cursor-pointer focus:outline-none">
-            <img
-              className="object-cover w-10 h-10 rounded-full"
-              src="https://github.com/achingbrain.png"
-              alt="username"
-            />
-            <div className="w-full pb-2">
-              <div className="flex justify-between">
-                <span className="block ml-2 font-semibold text-gray-600">
-                  Alex
-                </span>
-                <span className="block ml-2 text-sm text-gray-600">
-                  50 minutes
-                </span>
-              </div>
-              <span className="block ml-2 text-sm text-gray-600">
-                Good night
-              </span>
-            </div>
-          </a>
-          <a className="flex items-center px-3 py-2 text-sm transition duration-150 ease-in-out border-b border-gray-300 cursor-pointer hover:bg-gray-100 focus:outline-none">
-            <img
-              className="object-cover w-10 h-10 rounded-full"
-              src="https://github.com/hannahhoward.png"
-              alt="username"
-            />
-            <div className="w-full pb-2">
-              <div className="flex justify-between">
-                <span className="block ml-2 font-semibold text-gray-600">
-                  Hannah
-                </span>
-                <span className="block ml-2 text-sm text-gray-600">6 hour</span>
-              </div>
-              <span className="block ml-2 text-sm text-gray-600">
-                Good Morning
-              </span>
-            </div>
-          </a>
-        </li>
-      </ul>
-    </div>
-  )
-}
