+++ js-peer/src/components/room-list.tsx
@@ -1,0 +1,91 @@
+import React from 'react';
+
+
+export function RoomList() {
+  return (
+    <div className="border-r border-gray-300 lg:col-span-1">
+      <div className="mx-3 my-3">
+        <div className="relative text-gray-600">
+          <span className="absolute inset-y-0 left-0 flex items-center pl-2">
+            <svg
+              fill="none"
+              stroke="currentColor"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              viewBox="0 0 24 24"
+              className="w-6 h-6 text-gray-300"
+            >
+              <path d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"></path>
+            </svg>
+          </span>
+          <input
+            type="search"
+            className="block w-full py-2 pl-10 bg-gray-100 rounded outline-none"
+            name="search"
+            placeholder="Search"
+            required />
+        </div>
+      </div>
+
+      <ul className="overflow-auto h-[32rem]">
+        <h2 className="my-2 mb-2 ml-2 text-lg text-gray-600">Chats</h2>
+        <li>
+          <a className="flex items-center px-3 py-2 text-sm transition duration-150 ease-in-out border-b border-gray-300 cursor-pointer hover:bg-gray-100 focus:outline-none">
+            <img
+              className="object-cover w-10 h-10 rounded-full"
+              src="https://github.com/2color.png"
+              alt="username" />
+            <div className="w-full pb-2">
+              <div className="flex justify-between">
+                <span className="block ml-2 font-semibold text-gray-600">
+                  Daniel
+                </span>
+                <span className="block ml-2 text-sm text-gray-600">
+                  25 minutes
+                </span>
+              </div>
+              <span className="block ml-2 text-sm text-gray-600">bye</span>
+            </div>
+          </a>
+          <a className="flex items-center px-3 py-2 text-sm transition duration-150 ease-in-out bg-gray-100 border-b border-gray-300 cursor-pointer focus:outline-none">
+            <img
+              className="object-cover w-10 h-10 rounded-full"
+              src="https://github.com/achingbrain.png"
+              alt="username" />
+            <div className="w-full pb-2">
+              <div className="flex justify-between">
+                <span className="block ml-2 font-semibold text-gray-600">
+                  Alex
+                </span>
+                <span className="block ml-2 text-sm text-gray-600">
+                  50 minutes
+                </span>
+              </div>
+              <span className="block ml-2 text-sm text-gray-600">
+                Good night
+              </span>
+            </div>
+          </a>
+          <a className="flex items-center px-3 py-2 text-sm transition duration-150 ease-in-out border-b border-gray-300 cursor-pointer hover:bg-gray-100 focus:outline-none">
+            <img
+              className="object-cover w-10 h-10 rounded-full"
+              src="https://github.com/hannahhoward.png"
+              alt="username" />
+            <div className="w-full pb-2">
+              <div className="flex justify-between">
+                <span className="block ml-2 font-semibold text-gray-600">
+                  Hannah
+                </span>
+                <span className="block ml-2 text-sm text-gray-600">6 hour</span>
+              </div>
+              <span className="block ml-2 text-sm text-gray-600">
+                Good Morning
+              </span>
+            </div>
+          </a>
+        </li>
+      </ul>
+    </div>
+  );
+}
