+++ js-peer/src/pages/index.tsx
@@ -106,8 +106,6 @@ export default function Home() {
     [libp2p, maddr],
   )
 
-  // handleConnectToMultiaddr
-
   const handleMultiaddrChange = useCallback(
     (e: React.ChangeEvent<HTMLInputElement>) => {
       setMultiaddr(e.target.value)
